requirements.txt
asgiref==3.2.10
Django==3.0.8
django-geojson==3.0.0
django-leaflet==0.27.0
jsonfield==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
djangorestframework==3.11.0